Change IList.Count to an int

+++ DataStructures/DataStructures/ICollection.cs
@@ -2,6 +2,6 @@
 {
     public interface ICollection
     {
-        long Count { get; }
+        int Count { get; }
     }
 }
